Rename secret file
