+++ datasets/utils/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa: F401
-from .secret import SecretFetcher
+from .secret_fetcher import SecretFetcher
